simple calendar get/put
